backend-python/pyproject.toml
[project]
name = "backend-python"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beanie>=2.0.1",
    "fastapi[standard]>=0.128.0",
    "google-genai>=1.62.0",
    "google-maps-routing>=0.8.0",
    "openai>=2.17.0",
    "polyline>=2.0.4",
    "pydantic-settings>=2.12.0",
    "pymongo>=4.16.0",
    "python-dotenv>=1.2.1",
    "shapely>=2.1.2",
    "uvicorn[standard]>=0.40.0",
]

[project.scripts]
app = "app.main:app"

[dependency-groups]
dev = [
    "asgi-lifespan>=2.1.0",
    "httpx>=0.27.0",
    "mongomock-motor>=0.0.36",
    "pytest>=9.0.2",
    "pytest-asyncio>=0.25.0",
    "pytest-cov>=7.0.0",
    "ruff>=0.14.13",
]
